# TraceQL: Improve regexp perf