--- a/pkg/traceql/ast_validate.go
+++ b/pkg/traceql/ast_validate.go
@@ -1,6 +1,9 @@
 package traceql
 
-import "fmt"
+import (
+	"fmt"
+	"regexp"
+)
 
 // unsupportedError is returned for traceql features that are not yet supported.
 type unsupportedError struct {
@@ -177,6 +180,14 @@ func (o *BinaryOperation) validate() error {
 
 	if !o.Op.binaryTypesValid(lhsT, rhsT) {
 		return fmt.Errorf("illegal operation for the given types: %s", o.String())
+	}
+
+	// if this is a regex operator confirm the RHS is a valid regex
+	if o.Op == OpRegex || o.Op == OpNotRegex {
+		_, err := regexp.Compile(o.RHS.String())
+		if err != nil {
+			return fmt.Errorf("invalid regex: %s", o.RHS.String())
+		}
 	}
 
 	// this condition may not be possible to hit since it's not parseable.