@joe-elliott 2023-11-14
>  We can't do it during parsing because it's hard (impossible?) to return errors through the YACC snippets. Validate is ok but not quite right. So I added an `init()` method to some AST elements. Not asking for it here, but aligning some ideas.

Yeah, I ran up against these same decisions while making the change. I considered some kind of `init()` but it seemed like a big change for such a focused improvement. I do like the idea though if we have other places where it's useful.